AnalyzerReleases.Unshipped.md: both sides append new, distinct analyzer
ids, so keep both -- master's AIDN100 (ParameterUpdateInPlaceAnalyzer) and
this branch's AIDN101-AIDN104 (PaperOptimizerAnalyzer), in id order.

+++ src/Classification/MultiLabel/MultiLabelClassifierBase.cs
@@ -619,7 +619,7 @@ public abstract partial class MultiLabelClassifierBase<T> : IMultiLabelClassifie
             }
         }
 
-        double avgGradient = parameters.Length > 0 ? totalGradient / (input.Rows * NumLabels * parameters.Length) : 0;
+        double avgGradient = parameters.Length > 0 ? totalGradient / ((double)input.Rows * NumLabels * parameters.Length) : 0;
         for (int i = 0; i < gradients.Length; i++)
         {
             gradients[i] = NumOps.FromDouble(avgGradient);
